refactor: expand MRMR_WARNING_FLAGS with stronger checks
Add -Wconversion -Wsign-conversion -Wshadow -Wnull-dereference
-Wdouble-promotion -Wimplicit-fallthrough plus GCC-only -Wlogical-op
and -Wduplicated-cond. Each catches a class of bug the existing
-Wall -Wextra -Werror baseline does not.

Mechanical fallout fixed across library and benchmark code. Every
change is a local annotation (static_cast); no semantics altered.

Library:
  attribute_information::num_values() now static_cast<T>(_pdf.size());
  T is constrained to the [0, 255] domain by the upstream
  static_assert so the narrowing is value-preserving.

  compute_mi() casts loop indices i, j to value_type at the
  marginal_probability() call sites.

  delimiter_ctype::make_table() casts ~space to mask when clearing
  the bit in the lookup table.

Benchmarks:
  Five `bench_view_*` histogram updates of the form
  `++scratch[col1[i] * k + col2[i]]` now cast the index expression
  to size_t so the implicit unsigned char -> int -> size_t chain
  becomes explicit.

Catch2 SYSTEM include:
  Catch2's INTERFACE_INCLUDE_DIRECTORIES are reassigned to
  INTERFACE_SYSTEM_INCLUDE_DIRECTORIES post-FetchContent_MakeAvailable,
  so warnings from Catch2's own headers (Clang's -Wdouble-promotion
  firing inside catch_matchers_impl.hpp) do not break -Werror builds.

Verified: ctest 43/43 PASS at release/debug/sanitize on gcc 14.3.1.
Clang covered by the CI matrix.

+++ test/bench_view_access.cpp
@@ -113,7 +113,7 @@ TEST_CASE("bench: histogram via sorted index indirection", "[!benchmark][view-ac
     meter.measure([&] {
       scratch.assign(16, 0);
       for (std::size_t i = 0; i < 10000; ++i) {
-        ++scratch[col1[i] * 4 + col2[i]];
+        ++scratch[static_cast<std::size_t>(col1[i] * 4 + col2[i])];
       }
       return scratch[0];
     });
@@ -125,7 +125,7 @@ TEST_CASE("bench: histogram via sorted index indirection", "[!benchmark][view-ac
       scratch.assign(16, 0);
       for (std::size_t i = 0; i < indices_sorted.size(); ++i) {
         std::size_t idx = indices_sorted[i];
-        ++scratch[col1[idx] * 4 + col2[idx]];
+        ++scratch[static_cast<std::size_t>(col1[idx] * 4 + col2[idx])];
       }
       return scratch[0];
     });
@@ -137,7 +137,7 @@ TEST_CASE("bench: histogram via sorted index indirection", "[!benchmark][view-ac
       scratch.assign(16, 0);
       for (std::size_t i = 0; i < indices_unsorted.size(); ++i) {
         std::size_t idx = indices_unsorted[i];
-        ++scratch[col1[idx] * 4 + col2[idx]];
+        ++scratch[static_cast<std::size_t>(col1[idx] * 4 + col2[idx])];
       }
       return scratch[0];
     });
@@ -166,7 +166,7 @@ TEST_CASE("bench: sorted vs unsorted at 100K instances", "[!benchmark][view-acce
     meter.measure([&] {
       scratch.assign(16, 0);
       for (std::size_t i = 0; i < 100000; ++i) {
-        ++scratch[col1[i] * 4 + col2[i]];
+        ++scratch[static_cast<std::size_t>(col1[i] * 4 + col2[i])];
       }
       return scratch[0];
     });
@@ -178,7 +178,7 @@ TEST_CASE("bench: sorted vs unsorted at 100K instances", "[!benchmark][view-acce
       scratch.assign(16, 0);
       for (std::size_t i = 0; i < indices_sorted.size(); ++i) {
         std::size_t idx = indices_sorted[i];
-        ++scratch[col1[idx] * 4 + col2[idx]];
+        ++scratch[static_cast<std::size_t>(col1[idx] * 4 + col2[idx])];
       }
       return scratch[0];
     });
@@ -190,7 +190,7 @@ TEST_CASE("bench: sorted vs unsorted at 100K instances", "[!benchmark][view-acce
       scratch.assign(16, 0);
       for (std::size_t i = 0; i < indices_unsorted.size(); ++i) {
         std::size_t idx = indices_unsorted[i];
-        ++scratch[col1[idx] * 4 + col2[idx]];
+        ++scratch[static_cast<std::size_t>(col1[idx] * 4 + col2[idx])];
       }
       return scratch[0];
     });
@@ -254,7 +254,7 @@ TEST_CASE("bench: high cardinality impact", "[!benchmark][view-access]") {
     meter.measure([&] {
       scratch.assign(2500, 0);
       for (std::size_t i = 0; i < 100000; ++i) {
-        ++scratch[col1[i] * 50 + col2[i]];
+        ++scratch[static_cast<std::size_t>(col1[i] * 50 + col2[i])];
       }
       return scratch[0];
     });
@@ -266,7 +266,7 @@ TEST_CASE("bench: high cardinality impact", "[!benchmark][view-access]") {
       scratch.assign(2500, 0);
       for (std::size_t i = 0; i < indices_sorted.size(); ++i) {
         std::size_t idx = indices_sorted[i];
-        ++scratch[col1[idx] * 50 + col2[idx]];
+        ++scratch[static_cast<std::size_t>(col1[idx] * 50 + col2[idx])];
       }
       return scratch[0];
     });
